Fetched book from DB and fixed bugs related to typo. Also, created custom exception handling classes.

+++ src/main/java/com/example/bookstore/repositories/AuthorRepo.java
@@ -1,7 +1,8 @@
 package com.example.bookstore.repositories;
 
 import com.example.bookstore.entities.Author;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepo extends JpaRepository<Author , Integer> {
+public interface AuthorRepo extends JpaRepository<Author , Long> {
 }
