+++ src/main/java/com/example/bookstore/repositories/BookRepo.java
@@ -3,5 +3,5 @@ package com.example.bookstore.repositories;
 import com.example.bookstore.entities.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepo extends JpaRepository<Book, Integer> {
+public interface BookRepo extends JpaRepository<Book, Long> {
 }
